requirements.txt
pyobjc
pyautogui
pillow
pytesseract
sentence-transformers
faiss-cpu
numpy
pyyaml
pandas
pyobjc-framework-Accessibility